renamed some stuff + cleaned up some suggestions
