--- TowerNinja/Assets/Scripts/Enemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a37ee69b5246f0c4984405d3fe65eef5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
